refactor: 투표 목록 조회(커서) 리팩토링
