+++ api/src/main/java/com/programmers/bucketback/domains/vote/api/dto/response/VoteGetByCursorResponse.java
@@ -1,15 +1,15 @@
 package com.programmers.bucketback.domains.vote.api.dto.response;
 
-import com.programmers.bucketback.domains.vote.model.response.VoteCursorSummary;
-import com.programmers.bucketback.domains.vote.model.response.VotesGetServiceResponse;
-
 import java.util.List;
+
+import com.programmers.bucketback.common.cursor.CursorSummary;
+import com.programmers.bucketback.domains.vote.model.VoteCursorSummary;
 
 public record VoteGetByCursorResponse(
 	String nextCursorId,
 	List<VoteCursorSummary> votes
 ) {
-	public static VoteGetByCursorResponse from(final VotesGetServiceResponse serviceResponse) {
-		return new VoteGetByCursorResponse(serviceResponse.nextCursorId(), serviceResponse.votes());
+	public static VoteGetByCursorResponse from(final CursorSummary<VoteCursorSummary> cursorSummary) {
+		return new VoteGetByCursorResponse(cursorSummary.nextCursorId(), cursorSummary.summaries());
 	}
 }
