+++ api/src/main/java/com/programmers/bucketback/domains/vote/application/VoteService.java
@@ -4,16 +4,17 @@ import org.springframework.stereotype.Service;
 
 import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.common.cursor.CursorPageParameters;
+import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.domains.vote.application.dto.request.VoteCreateServiceRequest;
-import com.programmers.bucketback.domains.vote.application.dto.response.VoteGetServiceResponse;
-import com.programmers.bucketback.domains.vote.application.dto.response.VotesGetServiceResponse;
 import com.programmers.bucketback.domains.vote.domain.Vote;
 import com.programmers.bucketback.domains.vote.implementation.VoteAppender;
 import com.programmers.bucketback.domains.vote.implementation.VoteManager;
 import com.programmers.bucketback.domains.vote.implementation.VoteReader;
 import com.programmers.bucketback.domains.vote.implementation.VoteRemover;
+import com.programmers.bucketback.domains.vote.model.VoteCursorSummary;
 import com.programmers.bucketback.domains.vote.model.request.VoteSortCondition;
 import com.programmers.bucketback.domains.vote.model.request.VoteStatusCondition;
+import com.programmers.bucketback.domains.vote.model.response.VoteGetServiceResponse;
 import com.programmers.bucketback.error.exception.BusinessException;
 import com.programmers.bucketback.error.exception.ErrorCode;
 import com.programmers.bucketback.global.util.MemberUtils;
@@ -77,7 +78,7 @@ public class VoteService {
 		return voteReader.read(voteId, memberId);
 	}
 
-	public VotesGetServiceResponse getVotesByCursor(
+	public CursorSummary<VoteCursorSummary> getVotesByCursor(
 		final Hobby hobby,
 		final VoteStatusCondition statusCondition,
 		final VoteSortCondition sortCondition,
