+++ domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteSummary.java
@@ -1,0 +1,11 @@
+package com.programmers.bucketback.domains.vote.model;
+
+import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
+
+public record VoteSummary(
+	VoteInfo voteInfo,
+	Long item1Id,
+	Long item2Id,
+	String cursorId
+) {
+}
