--- domain/src/main/java/com/programmers/bucketback/domains/vote/model/response/VoteSummary.java
@@ -1,9 +1,0 @@
-package com.programmers.bucketback.domains.vote.model.response;
-
-public record VoteSummary(
-	VoteInfo voteInfo,
-	Long item1Id,
-	Long item2Id,
-	String cursorId
-) {
-}
