--- domain/src/main/java/com/programmers/bucketback/domains/vote/model/response/VotesGetServiceResponse.java
@@ -1,9 +1,0 @@
-package com.programmers.bucketback.domains.vote.model.response;
-
-import java.util.List;
-
-public record VotesGetServiceResponse(
-	String nextCursorId,
-	List<VoteCursorSummary> votes
-) {
-}
